- install-ruby-env-to-local.bashにsymlinkのチェックを追加

+++ scripts/install-ruby-env-to-local.bash
@@ -16,12 +16,14 @@ set -e
 
 cd $RBENV_SRC_DIR_PATH
 ls bin/ | while read line; do
-  ln -is $RBENV_SRC_DIR_PATH/bin/$line $LOCAL_DIR_PATH/bin/$line
+  if [[ ! -L $LOCAL_DIR_PATH/bin/$line ]]; then
+    ln -is $RBENV_SRC_DIR_PATH/bin/$line $LOCAL_DIR_PATH/bin/$line
+  fi
 done
 
 # install ruby-build
 cd $RUBY_BUILD_SRC_DIR_PATH
 PREFIX=$LOCAL_DIR_PATH ./install.sh
 
-printf "\ncomplete\n"
+printf "complete\n"
 exit 0
